Refactor

+++ src/main/java/ru/coffeecoders/questbot/entities/Question.java
@@ -15,7 +15,7 @@ public class Question {
     @Id
     @Column(name = "question_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long questionId;
+    private int questionId;
 
     @Column(name = "question")
     private String question;
@@ -35,11 +35,11 @@ public class Question {
     @Column(name = "question_group")
     private String group;
 
-    public long getQuestionId() {
+    public int getQuestionId() {
         return questionId;
     }
 
-    public void setQuestionId(long questionId) {
+    public void setQuestionId(int questionId) {
         this.questionId = questionId;
     }
